Refactor and Simplify Live Client
