+++ Deepgram/Constants/RequestMethod.cs
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a MIT license that can be found in the LICENSE file.
 // SPDX-License-Identifier: MIT
 
-namespace Deepgram.Enums;
+namespace Deepgram.Constants;
 
 public enum RequestMethod
 {
