+++ Deepgram/GlobalUsings.cs
@@ -14,8 +14,6 @@ global using System.Threading.Channels;
 global using System.Web;
 global using Deepgram.Abstractions;
 global using Deepgram.Constants;
-global using Deepgram.DeepgramEventArgs;
-global using Deepgram.Enums;
 global using Deepgram.Logger;
 global using Deepgram.Utilities;
 global using Microsoft.Extensions.DependencyInjection;
