+++ Deepgram/Models/Live/v1/LiveType.cs
@@ -2,12 +2,12 @@
 // Use of this source code is governed by a MIT license that can be found in the LICENSE file.
 // SPDX-License-Identifier: MIT
 
-using Deepgram.Models.Live.v1;
+namespace Deepgram.Models.Live.v1;
 
-namespace Deepgram.DeepgramEventArgs;
-
-public class ResponseReceivedEventArgs(EventResponse response) : EventArgs
+public enum LiveType
 {
-    public EventResponse Response { get; set; } = response;
+    Metadata,
+    Results,
+    UtteranceEnd,
+    SpeechStarted,
 }
-
