+++ examples/streaming/Streaming.csproj
@@ -8,7 +8,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Deepgram.Net.SDK" Version="4.0.0-alpha98" />
+    <ProjectReference Include="..\..\Deepgram\Deepgram.csproj" />
   </ItemGroup>
   
   <ItemGroup>
